+++ cargo-insta/tests/functional/main.rs
@@ -72,6 +72,7 @@ mod binary;
 mod delete_pending;
 mod glob_filter;
 mod inline;
+mod inline_snapshot_trimming;
 mod nextest_doctest;
 mod test_workspace_source_path;
 mod unreferenced;
@@ -281,6 +282,7 @@ impl TestProject {
                 let path_str = path.to_str().map(|s| s.replace('\\', "/")).unwrap();
                 format!("{}{}", "  ".repeat(entry.depth()), path_str)
             })
+            .filter(|line| !line.is_empty())
             .chain(std::iter::once(String::new()))
             .collect::<Vec<_>>()
             .join("\n")
@@ -508,6 +510,7 @@ fn test_wrong_indent_force() {
     foo
     foo
     "#, @r#"
+
                 foo
                 foo
     "#);
@@ -538,8 +541,6 @@ fn test_wrong_indent_force() {
 
 #[test]
 fn test_matches_fully_linebreaks() {
-    // Until #563 merges, we should be OK with different leading newlines, even
-    // in exact / full match mode.
     let test_project = TestFiles::new()
         .add_cargo_toml("exact-match-inline")
         .add_file(
@@ -555,6 +556,8 @@ fn test_additional_linebreak() {
         "insta_tests__tests",
     )
     "#, @r#"
+
+
     (
         "name_foo",
         "insta_tests__tests",
@@ -566,8 +569,7 @@ fn test_additional_linebreak() {
         )
         .create_project();
 
-    // Confirm the test passes despite the indent
-    let output = test_project
+    assert!(&test_project
         .insta_cmd()
         .args([
             "test",
@@ -577,8 +579,9 @@ fn test_additional_linebreak() {
             "--nocapture",
         ])
         .output()
-        .unwrap();
-    assert!(&output.status.success());
+        .unwrap()
+        .status
+        .success());
 }
 
 #[test]
@@ -619,6 +622,7 @@ fn foo_always_missing() {
     // Check for the name clash error message
     assert!(error_output.contains("Insta snapshot name clash detected between 'foo_always_missing' and 'test_foo_always_missing' in 'snapshot_name_clash_test'. Rename one function."));
 }
+
 #[test]
 fn test_hidden_snapshots() {
     let test_project = TestFiles::new()