+++ insta/src/snapshot.rs
@@ -6,12 +6,13 @@ use crate::{
 use once_cell::sync::Lazy;
 use std::env;
 use std::error::Error;
+use std::fmt;
 use std::fs;
 use std::io::{BufRead, BufReader, Write};
 use std::path::{Path, PathBuf};
 use std::rc::Rc;
 use std::time::{SystemTime, UNIX_EPOCH};
-use std::{borrow::Cow, fmt};
+use std::{borrow::Cow, iter::once};
 
 static RUN_ID: Lazy<String> = Lazy::new(|| {
     if let Ok(run_id) = env::var("NEXTEST_RUN_ID") {
@@ -646,6 +647,11 @@ impl TextSnapshotContents {
         TextSnapshotContents { contents, kind }
     }
 
+    /// Matches another snapshot without any normalization
+    pub fn matches_fully(&self, other: &TextSnapshotContents) -> bool {
+        self.contents == other.contents
+    }
+
     /// Snapshot matches based on the latest format.
     pub fn matches_latest(&self, other: &Self) -> bool {
         self.to_string() == other.to_string()
@@ -653,30 +659,64 @@ impl TextSnapshotContents {
 
     pub fn matches_legacy(&self, other: &Self) -> bool {
         fn as_str_legacy(sc: &TextSnapshotContents) -> String {
+            // First do the standard normalization
             let out = sc.to_string();
+            // Legacy snapshots trim newlines at the start.
+            let out = out.trim_start_matches(['\r', '\n']);
             // Legacy inline snapshots have `---` at the start, so this strips that if
             // it exists.
             let out = match out.strip_prefix("---\n") {
-                Some(old_snapshot) => old_snapshot.to_string(),
+                Some(old_snapshot) => old_snapshot,
                 None => out,
             };
             match sc.kind {
-                TextSnapshotKind::Inline => legacy_inline_normalize(&out),
-                TextSnapshotKind::File => out,
+                TextSnapshotKind::Inline => legacy_inline_normalize(out),
+                TextSnapshotKind::File => out.to_string(),
             }
         }
         as_str_legacy(self) == as_str_legacy(other)
     }
 
+    /// Convert a literal snapshot value (i.e. the string inside the quotes,
+    /// from a rust file) to the value we retain in the struct. This is a small
+    /// change to the value: we remove the leading newline and coerce newlines
+    /// to `\n`. Otherwise, the value is retained unnormalized (generally we
+    /// want to retain unnormalized values so we can run `matches_fully` on
+    /// them)
+    pub(crate) fn from_inline_literal(contents: &str) -> Self {
+        // If it's a single line string, then we don't do anything.
+        if contents.trim_end().lines().count() <= 1 {
+            return Self::new(contents.trim_end().to_string(), TextSnapshotKind::Inline);
+        }
+
+        // If it's multiline, we trim the first line, which should be empty.
+        // (Possibly in the future we'll do the same for the final line too)
+        let lines = contents.lines().collect::<Vec<&str>>();
+        let (first, remainder) = lines.split_first().unwrap();
+        let snapshot = {
+            // If the first isn't empty, something is up — include the first line
+            // and print a warning.
+            if first != &"" {
+                elog!("{} {}{}{}\n{}",style("Multiline inline snapshot values should start and end with a newline.").yellow().bold()," The current value will fail to match in the future. Run `cargo insta test --force-update-snapshots` to rewrite snapshots. The existing value's first line is `", first, "`. Full value:", contents);
+                once(first)
+                    .chain(remainder.iter())
+                    .cloned()
+                    .collect::<Vec<&str>>()
+                    .join("\n")
+            } else {
+                remainder.join("\n")
+            }
+        };
+        Self::new(snapshot, TextSnapshotKind::Inline)
+    }
+
     fn normalize(&self) -> String {
         let kind_specific_normalization = match self.kind {
-            TextSnapshotKind::Inline => normalize_inline_snapshot(&self.contents),
+            TextSnapshotKind::Inline => normalize_inline(&self.contents),
             TextSnapshotKind::File => self.contents.clone(),
         };
         // Then this we do for both kinds
-        let out = kind_specific_normalization
-            .trim_start_matches(['\r', '\n'])
-            .trim_end();
+        let out = kind_specific_normalization.trim_end();
         out.replace("\r\n", "\n")
     }
 
@@ -813,11 +853,6 @@ fn leading_space(value: &str) -> String {
 fn min_indentation(snapshot: &str) -> String {
     let lines = snapshot.trim_end().lines();
 
-    if lines.clone().count() <= 1 {
-        // not a multi-line string
-        return "".into();
-    }
-
     lines
         .filter(|l| !l.is_empty())
         .map(leading_space)
@@ -825,14 +860,134 @@ fn min_indentation(snapshot: &str) -> String {
         .unwrap_or("".into())
 }
 
-/// Removes excess indentation, and changes newlines to \n.
-fn normalize_inline_snapshot(snapshot: &str) -> String {
+/// Normalize snapshot value, which we apply to both generated and literal
+/// snapshots. Remove excess indentation, excess ending whitespace and coerce
+/// newlines to `\n`.
+fn normalize_inline(snapshot: &str) -> String {
+    // If it's a single line string, then we don't do anything.
+    if snapshot.trim_end().lines().count() <= 1 {
+        return snapshot.trim_end().to_string();
+    }
+
     let indentation = min_indentation(snapshot);
     snapshot
         .lines()
         .map(|l| l.get(indentation.len()..).unwrap_or(""))
         .collect::<Vec<&str>>()
         .join("\n")
+}
+
+#[test]
+fn test_normalize_inline_snapshot() {
+    fn normalized_of_literal(snapshot: &str) -> String {
+        normalize_inline(&TextSnapshotContents::from_inline_literal(snapshot).contents)
+    }
+
+    use similar_asserts::assert_eq;
+    // here we do exact matching (rather than `assert_snapshot`) to ensure we're
+    // not incorporating the modifications that insta itself makes
+
+    assert_eq!(
+        normalized_of_literal(
+            "
+   1
+   2
+"
+        ),
+        "1
+2"
+    );
+
+    assert_eq!(
+        normalized_of_literal(
+            r#"
+            1
+    2
+    "#
+        ),
+        r"        1
+2
+"
+    );
+
+    assert_eq!(
+        normalized_of_literal(
+            "
+            1
+            2
+    "
+        ),
+        r"1
+2
+"
+    );
+
+    assert_eq!(
+        normalized_of_literal(
+            "
+   1
+   2
+"
+        ),
+        "1
+2"
+    );
+
+    assert_eq!(
+        normalized_of_literal(
+            "
+        a
+    "
+        ),
+        "        a"
+    );
+
+    assert_eq!(normalized_of_literal(""), "");
+
+    assert_eq!(
+        normalized_of_literal(
+            r#"
+    a
+    b
+c
+    "#
+        ),
+        "    a
+    b
+c
+    "
+    );
+
+    assert_eq!(
+        normalized_of_literal(
+            "
+a
+    "
+        ),
+        "a"
+    );
+
+    // This is a bit of a weird case, but because it's not a true multiline
+    // (which requires an opening and closing newline), we don't trim the
+    // indentation. Not terrible if this needs to change. The next test shows
+    // how a real multiline string is handled.
+    assert_eq!(
+        normalized_of_literal(
+            "
+    a"
+        ),
+        "    a"
+    );
+
+    // This test will pass but raise a warning, so we comment it out for the moment.
+    // assert_eq!(
+    //     normalized_of_literal(
+    //         "a
+    //   a"
+    //     ),
+    //     "a
+    //   a"
+    // );
 }
 
 /// Extracts the module and snapshot name from a snapshot path
@@ -925,18 +1080,16 @@ fn legacy_inline_normalize(frozen_value: &str) -> String {
 }
 
 #[test]
-fn test_snapshot_contents() {
+fn test_snapshot_contents_to_inline() {
     use similar_asserts::assert_eq;
     let snapshot_contents =
         TextSnapshotContents::new("testing".to_string(), TextSnapshotKind::Inline);
     assert_eq!(snapshot_contents.to_inline(""), r#""testing""#);
 
-    let t = &"
-a
-b"[1..];
     assert_eq!(
-        TextSnapshotContents::new(t.to_string(), TextSnapshotKind::Inline).to_inline(""),
+        TextSnapshotContents::new("\na\nb".to_string(), TextSnapshotKind::Inline).to_inline(""),
         r##"r"
+
 a
 b
 ""##
@@ -954,6 +1107,7 @@ b
         TextSnapshotContents::new("\n    a\n    b".to_string(), TextSnapshotKind::Inline)
             .to_inline(""),
         r##"r"
+
 a
 b
 ""##
@@ -963,6 +1117,7 @@ b
         TextSnapshotContents::new("\na\n\nb".to_string(), TextSnapshotKind::Inline)
             .to_inline("    "),
         r##"r"
+
     a
 
     b
@@ -970,8 +1125,33 @@ b
     );
 
     assert_eq!(
+        TextSnapshotContents::new(
+            "ab
+    "
+            .to_string(),
+            TextSnapshotKind::Inline
+        )
+        .to_inline(""),
+        r#""ab""#
+    );
+
+    assert_eq!(
+        TextSnapshotContents::new(
+            "    ab
+    "
+            .to_string(),
+            TextSnapshotKind::Inline
+        )
+        .to_inline(""),
+        r##""    ab""##
+    );
+
+    assert_eq!(
         TextSnapshotContents::new("\n    ab\n".to_string(), TextSnapshotKind::Inline).to_inline(""),
-        r##""ab""##
+        r##"r"
+
+ab
+""##
     );
 
     assert_eq!(
@@ -1028,118 +1208,96 @@ a
 }
 
 #[test]
-fn test_normalize_inline_snapshot() {
+fn test_min_indentation() {
     use similar_asserts::assert_eq;
-    // here we do exact matching (rather than `assert_snapshot`)
-    // to ensure we're not incorporating the modifications this library makes
     assert_eq!(
-        normalize_inline_snapshot(
+        min_indentation(
             r#"
    1
    2
    "#,
         ),
-        r###"
-1
-2
-"###
+        "   ".to_string()
     );
 
     assert_eq!(
-        normalize_inline_snapshot(
+        min_indentation(
             r#"
             1
     2"#
         ),
-        r###"
-        1
-2"###
+        "    ".to_string()
     );
 
     assert_eq!(
-        normalize_inline_snapshot(
+        min_indentation(
             r#"
             1
             2
     "#
         ),
-        r###"
-1
-2
-"###
+        "            ".to_string()
     );
 
     assert_eq!(
-        normalize_inline_snapshot(
+        min_indentation(
             r#"
    1
    2
 "#
         ),
-        r###"
-1
-2"###
+        "   ".to_string()
     );
 
     assert_eq!(
-        normalize_inline_snapshot(
+        min_indentation(
             r#"
         a
     "#
         ),
-        "
-a
-"
+        "        ".to_string()
     );
 
-    assert_eq!(normalize_inline_snapshot(""), "");
+    assert_eq!(min_indentation(""), "".to_string());
 
     assert_eq!(
-        normalize_inline_snapshot(
+        min_indentation(
             r#"
     a
     b
 c
     "#
         ),
-        r###"
-    a
-    b
-c
-    "###
+        "".to_string()
     );
 
     assert_eq!(
-        normalize_inline_snapshot(
+        min_indentation(
             r#"
 a
     "#
         ),
-        "
-a
-    "
+        "".to_string()
     );
 
     assert_eq!(
-        normalize_inline_snapshot(
+        min_indentation(
             "
     a"
         ),
-        "
-a"
+        "    ".to_string()
     );
 
     assert_eq!(
-        normalize_inline_snapshot(
+        min_indentation(
             r#"a
   a"#
         ),
-        r###"a
-  a"###
+        "".to_string()
     );
 
     assert_eq!(
-        normalize_inline_snapshot(
+        normalize_inline(
             r#"
 			1
 	2"#
@@ -1150,7 +1308,7 @@ a"
     );
 
     assert_eq!(
-        normalize_inline_snapshot(
+        normalize_inline(
             r#"
 	  	  1
 	  	  2
@@ -1164,24 +1322,8 @@ a"
 }
 
 #[test]
-fn test_min_indentation() {
+fn test_min_indentation_additional() {
     use similar_asserts::assert_eq;
-    let t = r#"
-   1
-   2
-    "#;
-    assert_eq!(min_indentation(t), "   ".to_string());
-
-    let t = r#"
-            1
-    2"#;
-    assert_eq!(min_indentation(t), "    ".to_string());
-
-    let t = r#"
-            1
-            2
-    "#;
-    assert_eq!(min_indentation(t), "            ".to_string());
 
     let t = r#"
    1
@@ -1205,12 +1347,11 @@ c
     assert_eq!(min_indentation(t), "".to_string());
 
     let t = r#"
-a
-    "#;
+a"#;
     assert_eq!(min_indentation(t), "".to_string());
 
-    let t = "
-    a";
+    let t = r#"
+    a"#;
     assert_eq!(min_indentation(t), "    ".to_string());
 
     let t = r#"a
@@ -1237,7 +1378,7 @@ a
 #[test]
 fn test_inline_snapshot_value_newline() {
     // https://github.com/mitsuhiko/insta/issues/39
-    assert_eq!(normalize_inline_snapshot("\n"), "");
+    assert_eq!(normalize_inline("\n"), "");
 }
 
 #[test]
@@ -1274,7 +1415,7 @@ fn test_ownership() {
 #[test]
 fn test_empty_lines() {
     assert_snapshot!(r#"single line should fit on a single line"#, @"single line should fit on a single line");
-    assert_snapshot!(r#"single line should fit on a single line, even if it's really really really really really really really really really long"#, @"single line should fit on a single line, even if it's really really really really really really really really really long");
+    assert_snapshot!(r##"single line should fit on a single line, even if it's really really really really really really really really really long"##, @"single line should fit on a single line, even if it's really really really really really really really really really long");
 
     assert_snapshot!(r#"multiline content starting on first line